housing-service/requirements.txt
flask
python-dotenv
psycopg2-binary
PyJWT
pytest
pytest-flask